aact-827: WIP Starting to add framework to create foreign keys ‘after’ updates.
This has become a priority because the tool we’ve been using to develop the schema diagram (pgModeler) doesn’t work with PostreSQL v11.1, so we need to find an alternative.  Was investigating schemaSpy, which will generate a nice diagram automatically if the relationships are defined.  We should have those relationships defined anyway - to help users interpret the database.

Problem is: indexes & constraints significantly slow down database loads & it seems we’d need to refactor the loads to circumvent errors raised due to foreign key constraints.